Made images fit the comments and fixed a couple of javascript errors.
