+++ src/Components/AddPost/AddPost.css
@@ -27,6 +27,7 @@
 }
 
 #photoPreview{
+    display: block;
     max-width: 100%;
     max-width: 400px;
 }
